+++ packages/core/src/cells/marker-cell.tsx
@@ -4,6 +4,11 @@ import type { BaseDrawArgs, DrawArgs, InternalCellRenderer, PrepResult } from ".
 import { drawCheckbox } from "../internal/data-grid/render/draw-checkbox.js";
 import type { SpriteVariant } from "../index.js";
 
+interface FullMarkerFn extends MarkerFn {
+    start: number;
+    end: number;
+}
+
 export const markerCellRenderer: InternalCellRenderer<MarkerCell> = {
     getAccessibilityString: c => c.row.toString(),
     kind: InnerGridCellKind.Marker,
@@ -15,31 +20,25 @@ export const markerCellRenderer: InternalCellRenderer<MarkerCell> = {
     onClick: e => {
         const { bounds, cell, posX: x, posY: y } = e;
 
+        if (cell.functions?.length === 0) return undefined;
+
+        const meta = cell.meta;
+
         // 计算边界时应该按照每一项的盒子来计算
-        const isOverHeaderMarkerfn = cell.functions?.find?.(
-            item => x >= item.start && x <= item.end && y >= 0 && y <= bounds.height
+        const isOverHeaderMarkerfn = cell.functions.find(
+            item => x >= (item as FullMarkerFn).start && x <= (item as FullMarkerFn).end && y >= 0 && y <= bounds.height
         );
 
         if (isOverHeaderMarkerfn !== undefined && isOverHeaderMarkerfn.type !== "number") {
-            if (isOverHeaderMarkerfn.type === "expand" && cell?.node !== undefined) {
-                const { node } = cell;
+            let disabled = isOverHeaderMarkerfn?.disabled;
 
-                // eslint-disable-next-line @typescript-eslint/strict-boolean-expressions
-                node.collapsed = !node.collapsed;
-            } else {
-                let disabled = isOverHeaderMarkerfn?.disabled;
-
-                if (typeof isOverHeaderMarkerfn?.disabled === "function") {
-                    disabled = isOverHeaderMarkerfn?.disabled?.(cell.node);
-                }
-
-                if (disabled !== true) {
-                    isOverHeaderMarkerfn?.onClick?.(cell.node);
-                }
+            if (typeof isOverHeaderMarkerfn?.disabled === "function") {
+                disabled = isOverHeaderMarkerfn?.disabled?.(meta?.node);
             }
 
-            // 索引单元格可能会有多个功能区，每个功能区点击事件不一致。所以需要在调用onCelledited前设定所点击的功能区
-            cell.activeFn = isOverHeaderMarkerfn;
+            if (disabled !== true) {
+                isOverHeaderMarkerfn?.onClick?.(meta?.node);
+            }
 
             return cell;
         }
@@ -52,7 +51,7 @@ export const markerCellRenderer: InternalCellRenderer<MarkerCell> = {
 
         // 计算边界时应该按照每一项的盒子来计算
         const isOverHeaderMarkerfn = cell.functions?.find?.(
-            item => x >= item.start && x <= item.end && y >= 0 && y <= bounds.height
+            item => x >= (item as FullMarkerFn).start && x <= (item as FullMarkerFn).end && y >= 0 && y <= bounds.height
         );
 
         if (isOverHeaderMarkerfn && isOverHeaderMarkerfn.type !== "number") {
@@ -83,7 +82,7 @@ function deprepMarkerRowCell(args: Pick<BaseDrawArgs, "ctx">) {
 
 function drawMarkerRowCell(args: DrawArgs<MarkerCell>, cell: MarkerCell) {
     const { ctx, rect, hoverAmount, theme, spriteManager, hoverX = -100, highlighted, hoverY } = args;
-    const { row: index, markerKind, node, functions, checked, drawHandle, checkboxStyle } = cell;
+    const { row: index, markerKind, functions, checked, drawHandle, checkboxStyle, meta } = cell;
     const { x, y, width, height } = rect;
     const text = index.toString();
     const markerFontStyle = theme.markerFontFull;
@@ -96,7 +95,7 @@ function drawMarkerRowCell(args: DrawArgs<MarkerCell>, cell: MarkerCell) {
         if (markerKind === "both" && hoverAmount !== 0) {
             // ctx.globalAlpha = 1 - hoverAmount;
         }
-        ctx.fillStyle = highlighted ? accentColor ?? theme.markerTextAccent : color ?? theme.markerTextLight;
+        ctx.fillStyle = highlighted ? (accentColor ?? theme.markerTextAccent) : (color ?? theme.markerTextLight);
         ctx.font = `${hoverAmount > 0 ? "bold " + markerFontStyle : markerFontStyle}`;
         ctx.fillText(
             text,
@@ -110,67 +109,83 @@ function drawMarkerRowCell(args: DrawArgs<MarkerCell>, cell: MarkerCell) {
         }
     };
 
-    const drawExpand = (start: number, size: number, color?: string) => {
-        if (node !== undefined) {
-            const { children, collapsed, pid, isLast } = node;
+    const drawExpand = (item: MarkerFn, iconSize: number) => {
+        const { start } = item as FullMarkerFn;
 
-            ctx.save();
+        if (meta !== undefined) {
+            const { canOpen, isExpanded, isLeaf, depth, isLast } = meta;
+            // --- 绘制图标 (Expand/Collapse) ---
+            if (canOpen === true) {
+                // 计算鼠标是否悬停在图标上 (用于高亮或交互提示)
+                let isIconHovered = false;
+                if (
+                    hoverX !== undefined &&
+                    hoverX + rect.x > start &&
+                    hoverX + rect.x <= start + iconSize &&
+                    hoverY !== undefined &&
+                    hoverY > theme.cellVerticalPadding &&
+                    hoverY <= rect.height - theme.cellVerticalPadding
+                ) {
+                    isIconHovered = true;
+                }
 
-            if (children?.length) {
-                // 父元素，绘制expand icon
+                const size = isIconHovered ? iconSize + 2 : iconSize;
                 const startY = rect.y + (rect.height - size) / 2;
 
+                ctx.save();
                 spriteManager.drawSprite(
-                    collapsed === true ? "expand" : "collapse",
+                    isExpanded === true ? "collapse" : "expand",
                     "normal",
                     ctx,
                     start,
                     startY,
                     size,
-                    theme,
-                    1,
-                    size,
-                    color
+                    theme
                 );
-            } else if (pid !== undefined) {
-                // 子元素绘制刻度线
+                ctx.restore();
+
+                if (isIconHovered) {
+                    args.overrideCursor?.("pointer");
+                }
+            } else if (depth !== undefined && depth > 0 && isLeaf === true) {
+                // --- 绘制连接线  ---
+                ctx.save();
                 ctx.fillStyle = "none";
                 ctx.lineCap = "round";
                 ctx.lineJoin = "round";
                 ctx.lineWidth = 1.5;
                 ctx.strokeStyle = theme.markLine;
                 ctx.beginPath();
-                const startX = start + size / 2;
 
+                // 调整线的位置
+                const lineX = start + iconSize / 2;
+
+                // 绘制 L 型线
+                ctx.moveTo(lineX, y);
                 if (isLast === true) {
-                    ctx.moveTo(startX, y);
-                    ctx.lineTo(startX, y + height / 2);
-                    ctx.lineTo(startX + 6, y + height / 2);
+                    ctx.lineTo(lineX, y + height / 2); // 到底部中间
                 } else {
-                    ctx.moveTo(startX, y);
-                    ctx.lineTo(startX, y + height);
-                    ctx.moveTo(startX, y + height / 2);
-                    ctx.lineTo(startX + 6, y + height / 2);
+                    ctx.lineTo(lineX, y + height); // 到底部
                 }
+                // 横向短线
+                ctx.moveTo(lineX, y + height / 2);
+                ctx.lineTo(lineX + 6, y + height / 2);
 
                 ctx.stroke();
+                ctx.restore();
             }
-
-            ctx.restore();
-
-            return true;
         }
     };
 
     const drawIcon = (item: MarkerFn, totalWidth: number, size: number, disabled?: boolean) => {
-        const { content, start, color, hoverColor, rowDefKey, selectedColor, getColor } = item;
-        const icon = typeof content === "string" ? content : content?.(node);
+        const { content, start, end, color, hoverColor, selectedColor, selected } = item as FullMarkerFn;
+        const icon = typeof content === "string" ? content : content?.(meta?.node);
 
         let isHovered = false;
         if (
             hoverX !== undefined &&
-            hoverX > item.start &&
-            hoverX <= item.end &&
+            hoverX > start &&
+            hoverX <= end &&
             hoverY !== undefined &&
             hoverY > 0 &&
             hoverY <= rect.height
@@ -182,7 +197,7 @@ function drawMarkerRowCell(args: DrawArgs<MarkerCell>, cell: MarkerCell) {
 
         if (disabled === true) {
             variant = isHovered ? "disableHovered" : "disabled";
-        } else if (node !== undefined && rowDefKey !== undefined && node.rowDef?.[rowDefKey] === true) {
+        } else if ((typeof selected === "boolean" ? selected === true : selected?.(meta?.node)) ?? false) {
             variant = "selected";
         } else if (isHovered) {
             variant = isHovered ? "hovered" : "normal";
@@ -200,11 +215,15 @@ function drawMarkerRowCell(args: DrawArgs<MarkerCell>, cell: MarkerCell) {
                 theme,
                 1,
                 iconSize,
-                typeof color === "string" ? color : color?.(node),
+                typeof color === "string" ? color : color?.(meta?.node),
                 undefined,
-                typeof hoverColor === "string" ? hoverColor : hoverColor?.(node),
+                typeof hoverColor === "string" ? hoverColor : hoverColor?.(meta?.node),
                 undefined,
-                typeof selectedColor === "string" ? selectedColor : getColor?.("selected", node)
+                typeof selectedColor === "string"
+                    ? selectedColor
+                    : typeof color === "string"
+                      ? color
+                      : color?.(meta?.node, "selected")
             );
             if (hoverAmount !== 0) {
                 ctx.globalAlpha = 1;
@@ -213,11 +232,25 @@ function drawMarkerRowCell(args: DrawArgs<MarkerCell>, cell: MarkerCell) {
     };
 
     if (functions?.length) {
+        const visibleFunctions = functions.filter(fn => {
+            if (fn.visible === undefined) return true;
+            if (typeof fn.visible === "function") {
+                return fn.visible(meta?.node);
+            }
+            return fn.visible;
+        });
+
+        // 如果所有功能都不可见，直接返回
+        if (visibleFunctions.length === 0) {
+            return;
+        }
+
         let startX = x + padding;
+
         const perWidth = (rect.width - padding * 2) / functions.length;
 
-        functions
-            .sort((a, b) => a.order - b.order)
+        visibleFunctions
+            .sort((a, b) => (a.order ?? 0) - (b.order ?? 0))
             // eslint-disable-next-line unicorn/no-array-for-each
             .forEach(fnItem => {
                 const { spriteCbMap, type, size } = fnItem;
@@ -229,25 +262,30 @@ function drawMarkerRowCell(args: DrawArgs<MarkerCell>, cell: MarkerCell) {
                 }
 
                 const itemWidth =
-                    type === "number" ? textWith : type === "checkbox" ? size ?? 18 : size ?? theme.markerIconSize;
+                    type === "number"
+                        ? textWith
+                        : type === "checkbox"
+                          ? (size ?? theme.checkboxMaxSize)
+                          : (size ?? theme.markerIconSize);
 
-                if (functions.length === 1) {
+                if (visibleFunctions.length === 1) {
                     // 居中显示
-                    fnItem.start = rect.x + (rect.width - itemWidth) / 2;
-                    fnItem.end = fnItem.start + itemWidth;
+                    (fnItem as FullMarkerFn).start = rect.x + (rect.width - itemWidth) / 2;
+                    (fnItem as FullMarkerFn).end = (fnItem as FullMarkerFn).start + itemWidth;
                 } else {
-                    fnItem.start = startX;
-                    fnItem.end = startX + perWidth;
+                    (fnItem as FullMarkerFn).start = startX;
+                    (fnItem as FullMarkerFn).end = startX + perWidth;
 
-                    startX = fnItem.end;
+                    startX = (fnItem as FullMarkerFn).end;
                 }
 
                 const disabled =
                     fnItem.disabled === true ||
-                    (typeof fnItem.disabled === "function" && fnItem.disabled?.(node) === true);
+                    (typeof fnItem.disabled === "function" && fnItem.disabled?.(meta?.node) === true);
 
                 // 找出鼠标悬浮的项，修正鼠标悬浮样式
-                const isHovered = rectHoverX > fnItem.start && rectHoverX <= fnItem.end;
+                const isHovered =
+                    rectHoverX > (fnItem as FullMarkerFn).start && rectHoverX <= (fnItem as FullMarkerFn).end;
 
                 if (isHovered) {
                     if (fnItem.type === "number") {
@@ -260,10 +298,10 @@ function drawMarkerRowCell(args: DrawArgs<MarkerCell>, cell: MarkerCell) {
                 // eslint-disable-next-line unicorn/prefer-switch
                 if (type === "number") {
                     drawIndexNumber(
-                        fnItem.start,
+                        (fnItem as FullMarkerFn).start,
                         perWidth,
-                        typeof fnItem.color === "string" ? fnItem.color : fnItem.color?.(node),
-                        typeof fnItem.hoverColor === "string" ? fnItem.hoverColor : fnItem.hoverColor?.(node)
+                        typeof fnItem.color === "string" ? fnItem.color : fnItem.color?.(meta?.node),
+                        typeof fnItem.hoverColor === "string" ? fnItem.hoverColor : fnItem.hoverColor?.(meta?.node)
                     );
                 } else if (type === "checkbox") {
                     const offsetAmount = 7 * (checked ? hoverAmount : 1);
@@ -271,7 +309,7 @@ function drawMarkerRowCell(args: DrawArgs<MarkerCell>, cell: MarkerCell) {
                         ctx,
                         theme,
                         checked,
-                        drawHandle ? fnItem.start + offsetAmount : fnItem.start,
+                        drawHandle ? (fnItem as FullMarkerFn).start + offsetAmount : (fnItem as FullMarkerFn).start,
                         y,
                         drawHandle ? width - offsetAmount : width,
                         height,
@@ -283,11 +321,7 @@ function drawMarkerRowCell(args: DrawArgs<MarkerCell>, cell: MarkerCell) {
                         checkboxStyle
                     );
                 } else if (type === "expand") {
-                    drawExpand(
-                        fnItem.start,
-                        itemWidth,
-                        typeof fnItem.color === "string" ? fnItem.color : fnItem.color?.(node)
-                    );
+                    drawExpand(fnItem, itemWidth);
                 } else {
                     drawIcon(fnItem, perWidth, itemWidth, disabled);
                 }